requirements.txt
aws-cdk.core
aws-cdk.aws_ec2
security_groups